requirements.txt
# Core Dependencies for Goose + MiniMax Skills
# Generated from successful installation

# Core Data Processing
pandas>=3.0.0
numpy>=2.4.0
pillow>=11.3.0
requests>=2.33.0
matplotlib>=3.10.0

# Document Processing
pypdf>=6.9.0
python-docx>=1.2.0
openpyxl>=3.1.0
python-pptx>=1.0.0
"markitdown[pptx]">=0.1.5

# AI/ML Libraries
opencv-python>=4.13.0
scikit-image>=0.26.0
scikit-learn>=1.8.0
transformers>=5.5.0
torch>=2.11.0
torchvision>=0.26.0
torchaudio>=2.11.0

# Media Processing
moviepy>=2.1.0
pydub>=0.25.0
librosa>=0.11.0
soundfile>=0.13.0

# Web Frameworks
fastapi>=0.135.0
uvicorn>=0.43.0
streamlit>=1.56.0
gradio>=6.11.0
jinja2>=3.1.0
websockets>=16.0
aiofiles>=24.1.0

# Additional Utilities
python-dotenv>=1.2.0
pyyaml>=6.0.0
jsonschema>=4.26.0
click>=8.3.0
rich>=14.3.0
beautifulsoup4>=4.14.0
lxml>=6.0.0